Moved parse methods in spammer to utils, increased error catching coverage
